refactor network code 	modified:   .idea/misc.xml 	modified:   app/build.gradle 	modified:   app/src/main/AndroidManifest.xml 	new file:   app/src/main/java/com/hawksjamesf/simpleweather/AppComponent.java 	new file:   app/src/main/java/com/hawksjamesf/simpleweather/AppModule.java 	deleted:    app/src/main/java/com/hawksjamesf/simpleweather/HomeFragment.java 	new file:   app/src/main/java/com/hawksjamesf/simpleweather/MessageEvent.java 	modified:   app/src/main/java/com/hawksjamesf/simpleweather/SimpleWeatherApplication.java 	deleted:    app/src/main/java/com/hawksjamesf/simpleweather/WeatherAPIInterface.java 	deleted:    app/src/main/java/com/hawksjamesf/simpleweather/WebApiInterface.java 	new file:   app/src/main/java/com/hawksjamesf/simpleweather/bean/DataBean.java 	modified:   app/src/main/java/com/hawksjamesf/simpleweather/bean/RealTimeBean.java 	new file:   app/src/main/java/com/hawksjamesf/simpleweather/network/NetworkWrapper.java 	new file:   app/src/main/java/com/hawksjamesf/simpleweather/network/WeatherAPIInterface.java 	renamed:    app/src/main/java/com/hawksjamesf/simpleweather/FifteenDaysAdapter.java -> app/src/main/java/com/hawksjamesf/simpleweather/ui/FifteenDaysAdapter.java 	renamed:    app/src/main/java/com/hawksjamesf/simpleweather/FifteenDaysView.java -> app/src/main/java/com/hawksjamesf/simpleweather/ui/FifteenDaysView.java 	renamed:    app/src/main/java/com/hawksjamesf/simpleweather/HomeActivity.java -> app/src/main/java/com/hawksjamesf/simpleweather/ui/HomeActivity.java 	new file:   app/src/main/java/com/hawksjamesf/simpleweather/ui/HomeFragment.java 	renamed:    app/src/main/java/com/hawksjamesf/simpleweather/RefreshAdapter.java -> app/src/main/java/com/hawksjamesf/simpleweather/ui/RefreshAdapter.java 	renamed:    app/src/main/java/com/hawksjamesf/simpleweather/RefreshListView.java -> app/src/main/java/com/hawksjamesf/simpleweather/ui/RefreshListView.java 	new file:   app/src/main/java/com/hawksjamesf/simpleweather/ui/SetupWizardActivity.java 	new file:   app/src/main/java/com/hawksjamesf/simpleweather/ui/SplashAcvitivy.java 	renamed:    app/src/main/java/com/hawksjamesf/simpleweather/TemperatureView.java -> app/src/main/java/com/hawksjamesf/simpleweather/ui/TemperatureView.java 	new file:   app/src/main/res/layout/activity_setup_wizard.xml 	new file:   app/src/main/res/layout/activity_splash_acvitivy.xml 	new file:   app/src/main/res/layout/content_setup_wizard.xml 	new file:   app/src/main/res/layout/content_splash_acvitivy.xml 	modified:   app/src/main/res/layout/fragment_home.xml 	modified:   app/src/main/res/layout/item_fifteen_days_chart.xml 	modified:   app/src/main/res/layout/item_weather.xml 	modified:   app/src/main/res/values/dimens.xml 	modified:   app/src/main/res/values/strings.xml 	modified:   app/src/main/res/values/styles.xml 	modified:   build.gradle 	modified:   viewpagerindicator_library/build.gradle 	modified:   viewpagerindicator_library/viewpagerindicator_library.iml

+++ app/src/main/java/com/hawksjamesf/simpleweather/bean/RealTimeBean.java
@@ -6,7 +6,7 @@ import java.util.List;
  * Created by hawks.jamesf on 7/2/17.
  */
 
-public class RealTimeBean {
+public class RealTimeBean  {
 
 
     private String status;
@@ -75,7 +75,7 @@ public class RealTimeBean {
 
     public static class ResultBean {
         private String status;
-        private int temperature;
+        private float temperature;
         private String skycon;
         private double cloudrate;
         private int aqi;
@@ -93,11 +93,11 @@ public class RealTimeBean {
             this.status = status;
         }
 
-        public int getTemperature() {
+        public float getTemperature() {
             return temperature;
         }
 
-        public void setTemperature(int temperature) {
+        public void setTemperature(float temperature) {
             this.temperature = temperature;
         }
